Windows compilation part.
Did a few modifications to use MinGW in order to compile easily our code for the Windows platform on a Linux architecture. Here The Bootstrap file has been modified to use the MinGW's gcc. We had to add a toolchain-mingw.cmake to specify to cmake that we are compiling for Windows. The CMakeLists file has been updated to specify the correct path of used binaries.